Styling / download function
Add the cover flow to the index page. I still need to resize it. Had an
issue with the footer. One solution is to move footer to seperate file
then include after other view renders. added download function to routes
still need to test it. Also add table styling to roster page but does
not appear after logging in.